requirements/common.txt
Django>=1.6,<1.7
mysql-python
south
django-admin-bootstrapped
django-import-export

#partial dates
git+git://github.com/dracos/django-date-extensions.git
six
